+++ CHANGELOG.rst
@@ -5,6 +5,14 @@ The following is a summary of changes and improvements to
 :mod:`eulfedora`.  New features in each version should be listed, with
 any necessary information about installation or upgrade notes.
 
+1.4
+---
+
+* New **streaming** option for :class:`eulfedora.views.RawDatastreamView`
+  and :meth:`eulfedora.views.raw_datastream` to optionally return a
+  :class:`django.http.StreamingHttpResponse` (intended for use with
+  large datastream content).
+
 1.3.1
 -----